Refactored Forces to use new VertexUtility class
