--- FDG/frontend/src/forces/Repulsion.ts
@@ -1,39 +1,0 @@
-import type { Vertex } from "../classes/Vertex";
-import { PHYSICS } from "../constants";
-
-export function repulsion(
-    ctx: CanvasRenderingContext2D,
-    vertices: Array<Vertex>,
-    strength: number = PHYSICS.FORCES.COLOUMBS_LAW,
-    exponent: number = 1
-) {
-    for (let i = 0; i < vertices.length; i++) {
-        for (let j = i + 1; j < vertices.length; j++) {
-            const vertexA = vertices[i];
-            const vertexB = vertices[j];
-
-            // Calculate distance
-            const dx = vertexB.pos.x - vertexA.pos.x;
-            const dy = vertexB.pos.y - vertexA.pos.y;
-            const centerDistance = Math.sqrt(dx * dx + dy * dy);
-
-            if (centerDistance === 0) continue; // Avoid division by zero
-
-            const width_offset =
-                vertexA.getBoxWidth(ctx) / 2 + vertexB.getBoxWidth(ctx) / 2;
-
-            const edgeDistance = Math.max(centerDistance - width_offset, 2);
-            const force = strength / edgeDistance ** exponent;
-
-            // Unit vector (direction)
-            const unitX = dx / centerDistance;
-            const unitY = dy / centerDistance;
-
-            // Apply equal and opposite forces
-            vertexA.vector.x -= unitX * (force / vertexA.mass);
-            vertexA.vector.y -= unitY * (force / vertexA.mass);
-            vertexB.vector.x += unitX * (force / vertexB.mass);
-            vertexB.vector.y += unitY * (force / vertexB.mass);
-        }
-    }
-}
