+++ FDG/frontend/src/utility/Forces/Attraction.ts
@@ -1,0 +1,61 @@
+import { PHYSICS } from "../../../constants";
+import { Vertex } from "../../graph/Vertex";
+import { Vec } from "../../graph/Vec";
+import { Edge } from "../../graph/Edge";
+import { VertexUtility } from "../VertexUtility";
+
+export class Attraction {
+    static centerAttraction(origins: Set<Vertex>, canvas: HTMLCanvasElement) {
+        const canvasCenter = new Vec(canvas.width / 2, canvas.height / 2);
+
+        origins.forEach((origin) => {
+            if (origin.selected) return;
+            const dx = canvasCenter.x - origin.pos.x;
+            const dy = canvasCenter.y - origin.pos.y;
+
+            origin.vector.x += dx * PHYSICS.FORCES.CENTRAL_SPRING;
+            origin.vector.y += dy * PHYSICS.FORCES.CENTRAL_SPRING;
+        });
+    }
+
+    static springAttraction(
+        ctx: CanvasRenderingContext2D,
+        edges: Array<Edge>,
+        strength: number = PHYSICS.FORCES.SPRING
+    ) {
+        for (const edge of edges) {
+            const vertexA = edge.sourceRef;
+            const vertexB = edge.targetRef;
+
+            // calculates distance between centers of the vertices
+            const dx = vertexB.pos.x - vertexA.pos.x;
+            const dy = vertexB.pos.y - vertexA.pos.y;
+            const distance = Math.sqrt(dx * dx + dy * dy);
+
+            // avoid division by 0
+            if (distance === 0) {
+                continue;
+            }
+
+            // These are normalised directional vector components
+            const normX = dx / distance;
+            const normY = dy / distance;
+
+            const width_offset =
+                VertexUtility.getBoxWidth(ctx, vertexA) / 2 + VertexUtility.getBoxWidth(ctx, vertexB) / 2;
+
+            const displacement = distance - (PHYSICS.FORCES.REST_LENGTH + width_offset);
+            const force = strength * displacement;
+
+            // Pull two vertex's together
+            if (!vertexA.selected) {
+                vertexA.vector.x += (force / VertexUtility.getOriginalMass(vertexA)) * normX;
+                vertexA.vector.y += (force / VertexUtility.getOriginalMass(vertexA)) * normY;
+            }
+            if (!vertexB.selected) {
+                vertexB.vector.x -= (force / VertexUtility.getOriginalMass(vertexB)) * normX;
+                vertexB.vector.y -= (force / VertexUtility.getOriginalMass(vertexB)) * normY;
+            }
+        }
+    }
+}
